Maintain line numbers in rewritten code
This no longer uses UglifyJS to rewrite because it throws away line
numbers. Now being used is Narcissus which keeps line numbers around.
It's a bit harder to walk these ASTs but worth the line numbers.

No attempt is made to fix any existing deficiencies of the current
parser. This should be a bug-for-bug rewrite.

Also comments are maintained